"Amazon Clone: A e-commerce website built with HTML and CSS.
Amazon Clone: A e-commerce website built with HTML and CSS. This project aims to replicate the core features and user interface of Amazon, providing users with a familiar online shopping experience. The website includes key functionalities such as product listing, search and cart management. It is designed responsively to ensure optimal viewing across devices. This repository serves as a showcase for my front-end web development skills and can be used as a learning resource for those interested in building e-commerce websites using HTML and CSS.